# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: النبض المهني - Professional Pulse API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Register a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  # Skills
...
  /students/{id}/skills:
    get:
      operationId: getStudentSkills
      tags: [skills]
      summary: Get student skill profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentSkill"
    post:
      operationId: addStudentSkill
      tags: [skills]
      summary: Add or update a student skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentSkillInput"
      responses:
        "201":
          description: Skill added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentSkill"
  /students/{id}/skill-gap:
    get:
      operationId: getSkillGap
      tags: [skills]
      summary: Get skill gap analysis for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill gap analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGapAnalysis"
  /students/{id}/investment-value:
    get:
      operationId: getInvestmentValue
      tags: [skills]
      summary: Get professional investment value for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestmentValue"
  /students/{id}/career-paths:
    get:
      operationId: getCareerPaths
      tags: [skills]
      summary: Get recommended career paths for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Career path recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CareerPath"

  # Projects
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, studentId, major, year, overallScore, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        studentId:
          type: string
        major:
          type: string
        year:
          type: integer
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        overallScore:
          type: number
        readinessLevel:
          type: string
          enum: [beginner, developing, ready, advanced]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name, studentId, major, year]
      properties:
        name:
          type: string
        studentId:
          type: string
        major:
          type: string
        year:
          type: integer
        bio:
          type: string
        avatarUrl:
          type: string
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        major:
          type: string
        year:
          type: integer
        bio:
          type: string
        avatarUrl:
          type: string
...
    StudentSkill:
      type: object
      required: [id, studentId, skillId, skillName, category, level, verified, score]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        skillId:
          type: integer
        skillName:
          type: string
        category:
          type: string
        level:
          type: integer
          description: 1-5 proficiency level
        verified:
          type: boolean
        score:
          type: number
        verifiedAt:
          type: ["string", "null"]
    StudentSkillInput:
      type: object
      required: [skillId, level]
      properties:
        skillId:
          type: integer
        level:
          type: integer
    SkillGapAnalysis:
      type: object
      required: [studentId, overallReadiness, gaps, strengths]
      properties:
        studentId:
          type: integer
        overallReadiness:
          type: number
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/SkillGap"
        strengths:
          type: array
          items:
            type: string
    SkillGap:
      type: object
      required: [skillName, currentLevel, requiredLevel, gap, priority]
      properties:
        skillName:
          type: string
        currentLevel:
          type: number
        requiredLevel:
          type: number
        gap:
          type: number
        priority:
          type: string
          enum: [low, medium, high, critical]
        recommendation:
          type: string
    InvestmentValue:
      type: object
      required: [studentId, hourlyRateMin, hourlyRateMax, hourlyRateEstimate, currency, rarityScore, demandScore]
      properties:
        studentId:
          type: integer
        hourlyRateMin:
          type: number
        hourlyRateMax:
          type: number
        hourlyRateEstimate:
          type: number
        currency:
          type: string
        rarityScore:
          type: number
        demandScore:
          type: number
        marketInsight:
          type: string
    CareerPath:
      type: object
      required: [sector, matchScore, roles, demandLevel]
      properties:
        sector:
          type: string
        matchScore:
          type: number
        roles:
          type: array
          items:
            type: string
        demandLevel:
          type: string
          enum: [low, medium, high, critical]
        salaryRange:
          type: string
        description:
          type: string
...