# lib/api-zod/src/generated/types/skillGap.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * النبض المهني - Professional Pulse API
 * OpenAPI spec version: 0.1.0
 */
import type { SkillGapPriority } from './skillGapPriority';

export interface SkillGap {
  skillName: string;
  currentLevel: number;
  requiredLevel: number;
  gap: number;
  priority: SkillGapPriority;
  recommendation?: string;
}