# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * النبض المهني - Professional Pulse API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentReadinessLevel } from './studentReadinessLevel';

export interface Student {
  id: number;
  name: string;
  studentId: string;
  major: string;
  year: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  overallScore: number;
  readinessLevel?: StudentReadinessLevel;
  createdAt: string;
}