# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * النبض المهني - Professional Pulse API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  studentId: string;
  major: string;
  year: number;
  bio?: string;
  avatarUrl?: string;
}